New version
The Top-3 accuracy score improved from 0.52 to 0.80 after applying data augmentation. This was achieved by augmenting the dataset with a second version of each FAQ, where synonyms and rephrased text were used. The algorithm then calculates the cosine similarity between the user's input and both the original and the augmented FAQ versions. It selects the version with the highest similarity score. Finally, the algorithm outputs the top 3 most relevant FAQs, but it presents them in their original, unaltered formulation.

This approach helps by increasing the chances of matching the input with different variations of the same FAQ.